AsteroidOS: Support screenshots
With this, AsteroidOS is 100% supported in its current state, and
Gadgetbridge will work as a replacement for AsteroidOSSync.

Something to note here that would be useful for those who git-blame me
later is that asteroid-btsyncd will only send the screenshot size as
little-endian, explicitly. This is not documented in their
documentation, and therefore liable to change.