Even Realities G1: Move Packet Sequence Generation to the shared space
Previously, each side of the glasses had it's own sequence number, after
longer periods of time, these numbers would go out of sync and the
glasses would crash because the values are expected to be the same. Now
there is a single sequence number, and packets that are sent to both
lenses will now share the same sequence number. The heartbeat runner is
also moved to the shared space so that the time outs of the BLE
connections do not get skewed.