Garmin: Initial attempt at supporting incoming FitDefinition/FitData Messages
Add the code to generate outgoing FitDefinition- and FitDataStatusMessage.
Add support for incoming FitDefinitionMessage and register handler via deviceEvent (commented for now to avoid leaking).
Refactor FitLocalMessageHandler to support both outgoing and incoming handling of FitDefinition/FitData.

Most of the changes were not tested for the lack of hardware, but at least they show no regressions.