Reconnection: Fix a bug with the reconnection delay and make it more aggressive
This makes two changes:

First, it fixes a bug where the reconnection delay was always increased before the first reconnection attempt which meant that the first attempt was always 8 seconds later instead of the intended 4 seconds.

Second, it adds a new delay field to the the abstract coordinator which defaults to 2 seconds, and can be overridden by devices to get a smaller delay. If there are multiple devices waiting for reconnect, the lowest delay will be used.